Update prolog and problog packages
